Moved DeviceAuthFragment into it's own file
